Implement Waste Collection System API with FastAPI, including user authentication, registration, and database integration. Added project structure, environment configuration, and security features. Removed legacy code and organized services, repositories, and controllers for improved maintainability.
