--- api/api/utils/security.py
@@ -1,29 +1,0 @@
-from passlib.context import CryptContext
-import os
-from dotenv import load_dotenv
-
-# Load environment variables
-load_dotenv()
-
-# Security configuration
-SECRET_KEY = os.getenv("SECRET_KEY")
-if not SECRET_KEY:
-    raise ValueError(
-        "No SECRET_KEY set in environment. "
-        "Please set it with a secure key. "
-        "You can generate one using: python -c 'import secrets; print(secrets.token_urlsafe(64))'"
-    )
-
-ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", "30"))
-
-# Password hashing
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    """Verify a password against its hash."""
-    return pwd_context.verify(plain_password, hashed_password)
-
-def get_password_hash(password: str) -> str:
-    """Generate password hash."""
-    return pwd_context.hash(password)
