--- api/src/__init__.py
@@ -1,1 +1,0 @@
-# This file makes the directory a Python package
