--- api/src/core/validators.py
@@ -1,23 +1,0 @@
-import re
-from typing import Tuple
-
-def validate_password(password: str) -> Tuple[bool, str]:
-    """
-    Validates password strength and returns (is_valid, error_message)
-    """
-    if len(password) < 8:
-        return False, "Password must be at least 8 characters long"
-    
-    if not re.search(r"[A-Z]", password):
-        return False, "Password must contain at least one uppercase letter"
-    
-    if not re.search(r"[a-z]", password):
-        return False, "Password must contain at least one lowercase letter"
-    
-    if not re.search(r"\d", password):
-        return False, "Password must contain at least one number"
-    
-    if not re.search(r"[!@#$%^&*(),.?\":{}|<>]", password):
-        return False, "Password must contain at least one special character"
-    
-    return True, ""
