--- api/src/database/db.py
@@ -1,37 +1,0 @@
-from sqlalchemy import create_engine
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
-import os
-from dotenv import load_dotenv
-
-# Load environment variables from .env file
-load_dotenv()
-
-# Get database URL from environment variable
-DATABASE_URL = os.getenv('CONNECTION_URL')
-if not DATABASE_URL:
-    raise ValueError("CONNECTION_URL environment variable is not set")
-
-# Create SQLAlchemy engine
-engine = create_engine(DATABASE_URL)
-
-# Test connection
-try:
-    with engine.connect():
-        print("✅ Database connection successful!")
-except Exception as e:
-    print(f"❌ Database connection failed: {str(e)}")
-
-# Create session factory
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-# Create base class for models
-Base = declarative_base()
-
-# Dependency to get DB session
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
