+++ api/src/utils/security.py
@@ -1,26 +1,35 @@
+import os
 from datetime import datetime, timedelta
 from typing import Optional
 
+from dotenv import load_dotenv
 from jose import JWTError, jwt
 from passlib.context import CryptContext
 
-# Security configuration
-SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
-ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 30
+# Load environment variables
+load_dotenv()
 
+# Security configuration
+SECRET_KEY = os.getenv("SECRET_KEY", "")
+if not SECRET_KEY:
+    raise ValueError(
+        "No SECRET_KEY set in environment. "
+        "Please set it with a secure key. "
+        "You can generate one using: python -c 'import secrets; print(secrets.token_urlsafe(64))'"
+    )
+
+ALGORITHM = "HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", "30"))
+
+# Password hashing
 pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
 def verify_password(plain_password: str, hashed_password: str) -> bool:
-    """
-    Verify a plain password against a hashed password
-    """
+    """Verify a password against its hash."""
     return pwd_context.verify(plain_password, hashed_password)
 
 def get_password_hash(password: str) -> str:
-    """
-    Hash a password for storing
-    """
+    """Generate password hash."""
     return pwd_context.hash(password)
 
 def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
@@ -28,21 +37,16 @@ def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -
     Create a new access token
     """
     to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(minutes=15)
+    expire = datetime.utcnow() + (expires_delta if expires_delta else timedelta(minutes=15))
     to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
+    return jwt.encode(to_encode, str(SECRET_KEY), algorithm=ALGORITHM)
 
 def decode_access_token(token: str) -> Optional[dict]:
     """
     Decode an access token and return the payload if valid
     """
     try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        return payload
+        return jwt.decode(token, str(SECRET_KEY), algorithms=[ALGORITHM])
     except JWTError:
         return None
 
@@ -53,15 +57,13 @@ def create_refresh_token(data: dict) -> str:
     to_encode = data.copy()
     expire = datetime.utcnow() + timedelta(days=7)  # 7 days expiration for refresh token
     to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
+    return jwt.encode(to_encode, str(SECRET_KEY), algorithm=ALGORITHM)
 
 def decode_refresh_token(token: str) -> Optional[dict]:
     """
     Decode a refresh token and return the payload if valid
     """
     try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        return payload
+        return jwt.decode(token, str(SECRET_KEY), algorithms=[ALGORITHM])
     except JWTError:
-        return None 
+        return None
